+++ apps/webapp/src/data/wagmi/config/config.default.ts
@@ -1,5 +1,5 @@
 import { connectorsForWallets } from '@rainbow-me/rainbowkit';
-import { createConfig, http } from 'wagmi';
+import { createConfig, createStorage, http, noopStorage } from 'wagmi';
 import { mainnet, base, sepolia, arbitrum } from 'wagmi/chains';
 import {
   safeWallet,
@@ -17,6 +17,7 @@ import {
   TENDERLY_ARBITRUM_RPC_URL,
   TENDERLY_ARBITRUM_CHAIN_ID
 } from './testTenderlyChain';
+import { isTestnetId } from '@jetstreamgg/utils';
 
 export const tenderly = {
   id: TENDERLY_CHAIN_ID,
@@ -40,7 +41,7 @@ export const tenderly = {
 
 export const tenderlyBase = {
   id: TENDERLY_BASE_CHAIN_ID,
-  name: 'base_oct_9_0',
+  name: 'new-base-testnet-jan-27',
   network: 'tenderly base',
   // This is used by RainbowKit to display a chain icon for small screens. TODO: update to Base icon once available
   iconUrl: 'tokens/weth.svg',
@@ -110,15 +111,34 @@ export const wagmiConfigDev = createConfig({
     [sepolia.id]: http(import.meta.env.VITE_RPC_PROVIDER_SEPOLIA || ''),
     [tenderlyArbitrum.id]: http(import.meta.env.VITE_RPC_PROVIDER_TENDERLY_ARBITRUM || '')
   },
-  multiInjectedProviderDiscovery: false
+  multiInjectedProviderDiscovery: false,
+  storage: createStorage({
+    storage: typeof window !== 'undefined' && window.localStorage ? window.localStorage : noopStorage,
+    key: 'wagmi-dev'
+  })
 });
 
 export const wagmiConfigMainnet = createConfig({
-  chains: [mainnet, base],
+  chains: [mainnet, base, arbitrum],
   connectors,
   transports: {
     [mainnet.id]: http(import.meta.env.VITE_RPC_PROVIDER_MAINNET || ''),
-    [base.id]: http(import.meta.env.VITE_RPC_PROVIDER_BASE || '')
+    [base.id]: http(import.meta.env.VITE_RPC_PROVIDER_BASE || ''),
+    [arbitrum.id]: http(import.meta.env.VITE_RPC_PROVIDER_ARBITRUM || '')
   },
   multiInjectedProviderDiscovery: false
 });
+
+export const getSupportedChainIds = (chainId: number) => {
+  if (isTestnetId(chainId)) {
+    return [tenderly.id, tenderlyBase.id, tenderlyArbitrum.id];
+  }
+  return [mainnet.id, base.id, arbitrum.id];
+};
+
+export const getMainnetChainName = (chainId: number) => {
+  if (isTestnetId(chainId)) {
+    return tenderly.name;
+  }
+  return mainnet.name;
+};